% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dstr}
\alias{dstr}
\title{Siontis et al. (2018)'s network meta-analysis data}
\format{A arm-based dataset with 29 rows and 5 variables}
\usage{
data(dstr)
}
\description{
\itemize{
  \item \code{study}: Study ID
  \item \code{treat}: Treatment
  \item \code{trt}: Numbered treatment (1:CCTA, 2:CMR, 3:exercise ECG, 4:SPECT-MPI, 5:standard care, 6:Stress Echo)
  \item \code{n}: Sample size
  \item \code{d}: Number of events
}
}
\references{
Siontis, G. C., Mavridis, D., Greenwood, J. P., et al. (2018). 
Outcomes of non-invasive diagnostic modalities for the detection of coronary artery disease: network meta-analysis of diagnostic randomised controlled trials.
\emph{BMJ.}
\strong{360}: k504.
}
\keyword{datasets}
