\docType{package}
\name{PINMA-package}
\alias{PINMA-package}
\title{The 'PINMA' package.}
\description{
Improved Methods for Constructing Prediction Intervals for Network Meta-Analysis.
}
\references{
Noma, H., Hamura, Y., Sugasawa, S. and Furukawa, T. A. (2022+).
Improved methods to construct prediction intervals for network meta-analysis. Forthcoming.
}
